REN-46 : Refactoring images folder and fixing imports
