+++ frontend/src/components/UserHeader/index.tsx
@@ -1,7 +1,7 @@
 import DropdownMessage from "./DropdownMessage";
 import DropdownNotification from "./DropdownNotification";
 import DropdownUser from "./DropdownUser";
-import RentlyLogo from "../../assets/icons/rently-logo.svg?react";
+import RentlyLogo from '../../images/icon/rently-logo.svg?react';
 
 const Header = (props: {
   sidebarOpen: string | boolean | undefined;
